internal: Move the controllers directory to internal
