+++ internal/controllers/platformoperator_controller.go
